prototypes/web/fastapi_google_sso/requirements.txt
fastapi[standard]
fastapi-sso
python-jose[cryptography]